A working input box that doesn't look nice.
