+++ pokequiz/constants.py
@@ -3,12 +3,10 @@ import pygame
 pygame.init()
 WIDTH, HEIGHT = 800, 800
 
-# RED=(255,0,0)
-WHITE = (255, 255, 255)
-BLACK = (0, 0, 0)
-# BLUE=(0,0,255)
-CYAN = (52, 78, 91)
+WHITE = pygame.Color("white")
+BLACK = pygame.Color("black")
 
-BACKGROUND_COLOR = CYAN
-FONT = pygame.font.SysFont("arialblack", 40)
+BACKGROUND_COLOR = (52, 78, 91)
+FONT_HEIGHT = 40
+FONT = pygame.font.SysFont("arialblack", FONT_HEIGHT)
 TEXT_COLOR = WHITE
