+++ pokequiz/gui.py
@@ -1,0 +1,100 @@
+import pygame
+
+# from pokequiz.constants import WHITE, BLACK
+from pokequiz.constants import BLACK, FONT, WHITE
+
+# WHITE = pygame.Color('white')
+# BLACK = pygame.Color('black')
+COLOR_INACTIVE = pygame.Color("lightskyblue3")
+COLOR_ACTIVE = pygame.Color("dodgerblue2")
+# FONT = pygame.font.SysFont("arialblack", FONT_HEIGHT)
+
+
+class Button:
+    # Modified from https://stackoverflow.com/questions/63762922/whats-the-most-convenient-way-to-add-buttons-in-pygame-with-text
+    def __init__(self, size, text, pos, bgColor=WHITE, textColor=BLACK, center_text=False):
+        self.pos = pos
+        self.size = size
+        self.text = text
+        self.center_text = center_text
+        self.bgColor = bgColor
+        self.textColor = textColor
+        # self.font = pygame.font.Font(pygame.font.get_default_font(), self.size[1])
+        # self.textSurf = self.font.render(f"{self.text}", True, self.textColor)
+        # self.button = pygame.Surface((self.size[0], self.size[1])).convert()
+        # self.button.fill(self.bgColor)
+
+    def render(self, window):
+        self.font = pygame.font.Font(pygame.font.get_default_font(), self.size[1])
+        self.textSurf = self.font.render(f"{self.text}", True, self.textColor)
+        self.button = pygame.Surface((self.size[0], self.size[1])).convert()
+        self.button.fill(self.bgColor)
+
+        window.blit(self.button, (self.pos[0], self.pos[1]))
+        if self.center_text:
+            text_width, _ = self.textSurf.get_size()
+            text_position = (self.pos[0] + (self.size[0] - text_width) // 2, self.pos[1] + 5)
+        else:
+            text_position = (self.pos[0] + 1, self.pos[1] + 5)
+        # window.blit(self.textSurf, (self.pos[0]+1, self.pos[1]+5))
+        window.blit(self.textSurf, text_position)
+
+    def clicked(self, events):
+        mousePos = pygame.mouse.get_pos()  #  get the mouse position
+        for event in events:
+            if self.button.get_rect(topleft=self.pos).collidepoint(mousePos[0], mousePos[1]):
+                if event.type == pygame.MOUSEBUTTONDOWN:
+                    return True
+        return False
+
+    def flip(self):
+        self.bgColor, self.textColor = self.textColor, self.bgColor
+
+
+class InputBox:
+    # Modified from https://stackoverflow.com/questions/46390231/how-can-i-create-a-text-input-box-with-pygame
+    def __init__(self, x, y, w, h, text="", auto_active=False):
+        self.rect = pygame.Rect(x, y, w, h)
+        self.active = auto_active
+        self.color = COLOR_ACTIVE if self.active else COLOR_INACTIVE
+        self.text = text
+        self.update()
+        # self.txt_surface = FONT.render(text, True, self.color)
+
+    def handle_event(self, event):
+        if event.type == pygame.MOUSEBUTTONDOWN:
+            # If the user clicked on the input_box rect.
+            if self.rect.collidepoint(event.pos):
+                # Toggle the active variable.
+                self.active = not self.active
+            else:
+                self.active = False
+            # Change the current color of the input box.
+            self.color = COLOR_ACTIVE if self.active else COLOR_INACTIVE
+        if event.type == pygame.KEYDOWN:
+            if self.active:
+                if event.key == pygame.K_RETURN:
+                    print(self.text)
+                    current_text = self.text
+                    self.text = ""
+                    self.update()
+                    return current_text
+                elif event.key == pygame.K_BACKSPACE:
+                    self.text = self.text[:-1]
+                else:
+                    self.text += event.unicode
+                # Re-render the text.
+                self.update()
+        return None
+
+    def update(self):
+        self.txt_surface = FONT.render(self.text, True, self.color)
+        # Resize the box if the text is too long.
+        width = max(200, self.txt_surface.get_width() + 10)
+        self.rect.w = width
+
+    def draw(self, screen):
+        # Blit the text.
+        screen.blit(self.txt_surface, (self.rect.x + 5, self.rect.y + 5))
+        # Blit the rect.
+        pygame.draw.rect(screen, self.color, self.rect, 2)
